+++ bfabric_scripts/src/bfabric_scripts/cli/login/logout.py
@@ -9,11 +9,10 @@ from typing import Annotated
 import cyclopts
 import yaml
 
-from bfabric._oauth._constants import DEFAULT_CLIENT_ID
-
 from bfabric.config import DEFAULT_CONFIG_FILE
 from bfabric.config.config_file import ConfigFile
 from bfabric._oauth.token_cache import TokenCache, compute_token_cache_path
+from bfabric_scripts.cli.login._constants import DEFAULT_CLIENT_ID
 
 
 def cmd_login_logout(